crewai/requirements.txt
crewai==1.14.2
crewai-tools==1.14.2
python-dotenv==1.1.1
